math linking problem

+++ math/sigmoid.c
@@ -1,25 +1,26 @@
 #include <stdio.h>
 #include <math.h>
+#include "sigmoid.h"
 
-double sigmoid(int x)
+double sigmoid(double x)
 {
-    return 1/(1+exp(-x));
+    return 1 / (1 + exp(-x));
 }
 
 double* vecd_sigmoid(double* vector, size_t len)
 {
     for(size_t i = 0; i < len; ++i)
     {
-        int tmp = *(vector +i);
+        double tmp = *(vector +i);
         *(vector+i) = sigmoid(tmp);
     }
     return vector;
 }
 
-double prime_sigmoid(int x)
+double prime_sigmoid(double x)
 {
     double sigx = sigmoid(x);
-    return sigx * (1- sigx);
+    return sigx * (1 - sigx);
 }
 
 
@@ -27,8 +28,9 @@ double* vecd_prime_sigmoid(double* vector, size_t len)
 {
     for(size_t i = 0; i < len; ++i)
     {
-        int tmp = *(vector +i);
+        double tmp = *(vector +i);
         *(vector+i) = prime_sigmoid(tmp);
     }
+
     return vector;
 }
