+++ math/sigmoid.h
@@ -1,12 +1,11 @@
 #pragma once
 #include <stddef.h>
 
-
-double sigmoid(int x);
+double sigmoid(double x);
 
 double* vecd_sigmoid(double* vector, size_t len);
 
-double prime_sigmoid(int x);
+double prime_sigmoid(double x);
 
 double* vecd_prime_sigmoid(double* vector, size_t len);
